% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sharepoint_basepath}
\alias{get_sharepoint_basepath}
\title{Returns the path to the SharePoint folder path}
\usage{
get_sharepoint_basepath()
}
\value{
character vector
}
\description{
Returns the path to the SharePoint folder path
}
