% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{yem_build_path}
\alias{yem_build_path}
\title{Build paths relative to SharePoint root}
\usage{
yem_build_path(..., must_exist = TRUE)
}
\arguments{
\item{...}{Character vectors, arguments to construct a path}

\item{must_exist}{Logical indicating if final path should exist}
}
\value{
A normalized path string starting from SharePoint root
}
\description{
Build paths relative to SharePoint root
}
