% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\docType{data}
\name{yem_core_deps}
\alias{yem_core_deps}
\title{Core Dependencies}
\format{
An object of class \code{character} of length 7.
}
\usage{
yem_core_deps
}
\description{
Core Dependencies
}
\keyword{datasets}
