% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\docType{data}
\name{yem_dev_deps}
\alias{yem_dev_deps}
\title{Development Dependencies}
\format{
An object of class \code{character} of length 7.
}
\usage{
yem_dev_deps
}
\description{
Development Dependencies
}
\keyword{datasets}
