% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{yem_find_file}
\alias{yem_find_file}
\title{Find files in SharePoint by exact name match}
\usage{
yem_find_file(filename)
}
\arguments{
\item{filename}{Character string of exact filename to find}
}
\value{
Character vector of matching relative paths
}
\description{
Find files in SharePoint by exact name match
}
