% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{yem_get_sharepoint_basepath}
\alias{yem_get_sharepoint_basepath}
\title{Returns the path to the SharePoint folder path}
\usage{
yem_get_sharepoint_basepath()
}
\value{
character vector representing the normalized SharePoint base path
}
\description{
Returns the path to the SharePoint folder path
}
