% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{yem_init_dependencies}
\alias{yem_init_dependencies}
\title{Initialize package dependencies}
\usage{
yem_init_dependencies(
  path,
  dependencies = character(),
  core_deps = yem_core_deps
)
}
\arguments{
\item{path}{Path to package root}

\item{dependencies}{Character vector of additional dependencies}

\item{core_deps}{Character vector of core dependencies to include}
}
\description{
Initialize package dependencies
}
\keyword{internal}
