% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{yem_init_documentation}
\alias{yem_init_documentation}
\title{Setup documentation structure}
\usage{
yem_init_documentation(path)
}
\arguments{
\item{path}{Path to package root}
}
\description{
Setup documentation structure
}
\keyword{internal}
