% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{yem_init_folders}
\alias{yem_init_folders}
\title{Create standard folder structure}
\usage{
yem_init_folders(path)
}
\arguments{
\item{path}{Path to package root}
}
\description{
Create standard folder structure
}
\keyword{internal}
