% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{yem_init_project}
\alias{yem_init_project}
\title{Initialize analysis project}
\usage{
yem_init_project(project_name, project_dir = ".", use_git = TRUE)
}
\arguments{
\item{project_name}{chracter vector project name}

\item{project_dir}{path}

\item{use_git}{boolean Whether to use git or not}
}
\value{
boolean invisible
}
\description{
Initialize analysis project
}
