% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{yem_init_targets}
\alias{yem_init_targets}
\title{Setup targets pipeline template}
\usage{
yem_init_targets(path)
}
\arguments{
\item{path}{Path to package root}
}
\description{
Setup targets pipeline template
}
\keyword{internal}
