% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{yem_init_project}
\alias{yem_init_project}
\title{Initialize a new data analysis project}
\usage{
yem_init_project(name, path = ".", git = TRUE)
}
\arguments{
\item{name}{Name of the project/package}

\item{path}{Path where to create the project}

\item{git}{Should git be initialized}
}
\description{
Initialize a new data analysis project
}
